Update formatting, null checks, and loop guards
- Refined `.clang-format` with explicit include categories, enabled short loops/functions on a single line, and sorted includes.
- Added a null check for `renderDevice` in `DimPaperSpaceOverlay` to prevent possible null dereference.
- Corrected `EoDbLineType::IndexToString()` to use signed formatting and explicit cast.
- Improved robustness in `DisplayFilAreaHatch` by skipping processing when `iStrs <= 0`.
- Made boolean arithmetic explicit in `absmatcher.h` for type safety in `eol()`.

+++ Application/AeSys/EoDbLineType.h
@@ -20,7 +20,7 @@ class EoDbLineType : public CObject {
 
   CString IndexToString() const {
     CString indexAsString;
-    indexAsString.Format(L"%03u", m_Index);
+    indexAsString.Format(L"%03d", static_cast<int>(m_Index));
     return indexAsString;
   }
